fixed small maps bugs
